requirements.txt
yara-python
capstone
keystone-engine
angr
requests
